Allow mocking and GraphQL query.
